House Cleaning
- Updated / Renamed files to pre-emptively fix issues with Heroku and its case issues
