+++ controllers/api/index.js
@@ -1,7 +1,7 @@
-const router = require('express').Router();
-const userRoutes = require('./userRoutes');
-const genreRoutes = require('./genreRoutes');
+const router = require("express").Router();
+const userRoutes = require("./user_routes");
+const genreRoutes = require("./genre_routes");
 
-router.use('/users', userRoutes);
-router.use('/genres', genreRoutes);
+router.use("/users", userRoutes);
+router.use("/genres", genreRoutes);
 module.exports = router;
