+++ controllers/api/user_routes.js
@@ -1,0 +1,49 @@
+const router = require("express").Router();
+const {} = require("../../models");
+
+// Login
+router.post("/login", async (req, res) => {
+  try {
+    const userData = await User.findOne({
+      where: {
+        username: req.body.username,
+      },
+    });
+
+    if (!userData) {
+      res.status(400).json({ message: "No user with that username!" });
+      return;
+    }
+
+    const validPassword = await userData.checkPassword(req.body.password);
+
+    if (!validPassword) {
+      res.status(400).json({ message: "Incorrect password!" });
+      return;
+    }
+
+    req.session.save(() => {
+      req.session.logged_in = true;
+
+      res
+        .status(200)
+        .json({ user: userData, message: "You are now logged in!" });
+    });
+  } catch (err) {
+    console.log(err);
+    res.status(500).json(err);
+  }
+});
+
+// Logout
+router.post("/logout", (req, res) => {
+  if (req.session.logged_in) {
+    req.session.destroy(() => {
+      res.status(204).end();
+    });
+  } else {
+    res.status(404).end();
+  }
+});
+
+module.exports = router;
