+++ public/css/style.css
@@ -1,0 +1,18 @@
+body,
+html {
+  height: 100%;
+}
+
+#bg-img-1 {
+  /* The image used */
+  background-image: url("../images/homeimage1.jpg");
+
+  /* Full height */
+  height: 100%;
+
+  /* Create the parallax scrolling effect */
+  background-attachment: fixed;
+  background-position: center;
+  background-repeat: no-repeat;
+  background-size: cover;
+}
