+++ views/home.handlebars
@@ -1,5 +1,7 @@
-<div class="d-flex justify-content-center">
-  <p class="col-md-8 fs-5 text-center">The ONLY place on the web where you can
-    customize, store and share your musical rigs! Sign up for FREE and get
-    started right away!</p>
-</div>
+<section id="bg-img-1" class="">
+  <div class="d-flex justify-content-center">
+    <p class="col-md-8 fs-5 text-center">The ONLY place on the web where you can
+      customize, store and share your musical rigs! Sign up for FREE and get
+      started right away!</p>
+  </div>
+</section>
