+++ views/layouts/main.handlebars
@@ -26,7 +26,7 @@
               <div>
                 <img
                   src="/images/bg1.png"
-                  class="vw-100 h-auto flex-fill p-0"
+                  class="vw-100 h-auto p-0"
                   alt="Spackled wall texture with vector drawings of a saxophone, drum kit, electric guitar, a pair of maracas and the Rig Match Me logo in the center"
                 />
               </div>
@@ -62,14 +62,14 @@
                   <li class="nav-item">
                     <a
                       class="nav-link text-light"
-                      href="/newRig"
+                      href="/new_rig"
                       id="newRig"
                     ><h4>Build a New Rig</h4></a>
                   </li>
                   <li class="nav-item">
                     <a
                       class="nav-link text-light"
-                      href="/newRig"
+                      href="/new_rig"
                       id="newRig"
                     ><h4>View My Rigs</h4></a>
                   </li>
@@ -92,7 +92,7 @@
         {{! NAVBAR END}}
 
       </header>
-      <main class="container container-fluid mt-5">
+      <main>
         <!-- Render the sub layout -->
         {{{body}}}
       </main>
