+++ views/new_rig.handlebars
